e4p2_capstone/aws/text_Classify_functions/requirements.txt
requests_toolbelt
pandas
numpy==1.18.5
Pillow==7.2.0